Added Surviving Free Culture article
As the author of it, happy to add a reference to burnout.io at the end. Should burnout.io be the place that I recommend to continue discussion on such things at? Would be nice for all the discussion to be consolidated rather than being in the comments of several different blog posts.