Update actionables.md
Adding a couple of more actionables here, specifically fueling up:

Take micro-breaks and play:
Encouragement to take micro-breaks and play. After reading Charlie Hoehn's "Play It Away", I found that I didn't play, and wasn't really taking time to get away from work when at work. I found it amazingly helpful to get me out of the position of feeling hopeless and lost as far as making changes at work, and figure it may be helpful for other people too. 

The note section here is also something that I've found helpful, as my wife and I both have an unhealthy relationship with work. Having someone around who can directly see if you're bringing work home can help with burnout as well. I'm not sure if I need to cite this, but it's made a difference for me. 

Take up a hobby:
This was also a crucial step for me. I took a week off of work to learn how to home brew. The week of not worrying about work and the constant stream of requests made all the difference.

I realize that the stuff I've added isn't going to be possible or feasible for everyone, but they helped pull me out of a dark place--hopefully they can do the same for others! One other note--I'm conscious of the link to "Play It Away", and that it may not exactly be the best fit, but I leave that to your judgement.